# optimize transpose CUDA kernel