--- a/onnxruntime/core/providers/cuda/tensor/transpose.cc
+++ b/onnxruntime/core/providers/cuda/tensor/transpose.cc
@@ -98,20 +98,76 @@ Status Transpose::DoTranspose(const cublasHandle_t cublas_handle,
   const std::vector<int64_t>& output_dims = output.Shape().GetDims();
 
   auto rank = static_cast<int32_t>(input_dims.size());
-  TensorPitches original_input_strides(input_dims);
-  TensorPitches original_output_strides(output_dims);
 
-  TArray<int64_t> input_strides(rank);
-  for (auto i = 0; i < rank; i++) {
-    input_strides[i] = original_input_strides[permutations[i]];
-  }
-  TArray<fast_divmod> output_strides(rank);
-  for (auto i = 0; i < rank; i++) {
-    output_strides[i] = fast_divmod(gsl::narrow_cast<int>(original_output_strides[i]));
+  // flatten some continguous dimensions
+  auto new_rank = rank;
+  std::vector<size_t> new_permutations(permutations);
+  std::vector<int64_t> new_input_dims(input_dims);
+  std::vector<int64_t> new_output_dims(output_dims);
+  // for (auto i = rank - 1; i > 0; i--) {
+  //   auto curr = permutations[i];
+  //   auto prev = permutations[i - 1];
+  //   if (prev + 1 == curr) {
+  //     // all dims bigger than curr need to be reduced by 1 due to the merging.
+  //     for (auto j = 0; j < new_rank; j++) {
+  //       if (new_permutations[j] > curr) {
+  //         new_permutations[j] -= 1;
+  //       }
+  //     }
+  //     for (auto j = i+1; j < new_rank; j++) {
+  //       new_permutations[j-1] = new_permutations[j];
+  //     }
+
+  //     new_input_dims[prev] *= new_input_dims[curr];
+  //     new_input_dims[curr] = 1;
+  //     for (auto j = curr+1; j < new_rank; j++) {
+  //       new_input_dims[j-1] = new_input_dims[j];
+  //     }
+
+  //     new_output_dims[i-1] *= new_output_dims[i];
+  //     new_output_dims[i] = 1;
+  //     for (auto j = i+1; j < new_rank; j++) {
+  //       new_output_dims[j-1] = new_output_dims[j];
+  //     }
+  //     new_rank--;
+  //   }
+  // }
+
+  TensorPitches new_input_strides(new_input_dims);
+  TensorPitches new_output_strides(new_output_dims);
+
+  // Optimize for 4D tensor permutation
+  TArray<int64_t> input_shape(new_input_dims);
+  TArray<int64_t> tmp_input_strides(new_input_strides);
+
+  TArray<int64_t> tmp_output_strides(new_rank);
+  for (auto i = 0; i < new_rank; i++) {
+    tmp_output_strides[i] = new_output_strides[new_permutations[i]];
   }
 
   size_t element_size = input.DataType()->Size();
-  auto status = TransposeImpl(element_size, rank, input_strides, input.DataRaw(),
+  // last dim won't be changed.
+  if (rank == 4 && permutations[3] == 3) {
+    // each cuda thread will process the number of elements = 4 * sizeof(float) / element_size
+    int64_t num_elements = input_shape[2] * input_shape[3] / (4 * sizeof(float) / element_size);
+    // num_elements must be aligned with warp size: 32
+    if (num_elements <= kernel.GetDeviceProp().maxThreadsPerBlock && (num_elements & 0x1F) == 0) {
+      return Transpose4DImpl(element_size, input_shape, tmp_input_strides, input.DataRaw(),
+                             tmp_output_strides, output.MutableDataRaw(), output.Shape().Size());
+    }
+  }
+
+  TArray<int64_t> input_strides(new_rank);
+  for (auto i = 0; i < new_rank; i++) {
+    input_strides[i] = new_input_strides[new_permutations[i]];
+  }
+
+  TArray<fast_divmod> output_strides(new_rank);
+  for (auto i = 0; i < new_rank; i++) {
+    output_strides[i] = fast_divmod(gsl::narrow_cast<int>(new_output_strides[i]));
+  }
+
+  auto status = TransposeImpl(element_size, new_rank, input_strides, input.DataRaw(),
                               output_strides, output.MutableDataRaw(), output.Shape().Size());
 
   return status;