--- a/onnxruntime/core/providers/cuda/tensor/transpose.h
+++ b/onnxruntime/core/providers/cuda/tensor/transpose.h
@@ -23,7 +23,8 @@ class Transpose final : public CudaKernel, public TransposeBase {
                             const std::vector<size_t>& permutations, const Tensor& input, Tensor& output);
 
   //  `input_shape_override` (if provided) overrides the shape of `input` for compute purposes
-  static Status DoTranspose(const cublasHandle_t cublas_handle,
+  static Status DoTranspose(const cudaDeviceProp& prop,
+                            const cublasHandle_t cublas_handle,
                             const std::vector<size_t>& permutations,
                             const Tensor& input, Tensor& output, const TensorShape* input_shape_override = nullptr);
 };