--- a/onnxruntime/core/providers/cuda/tensor/transpose_impl.cu
+++ b/onnxruntime/core/providers/cuda/tensor/transpose_impl.cu
@@ -7,6 +7,80 @@
 namespace onnxruntime {
 namespace cuda {
 
+constexpr int TILE_DIM = 32;
+constexpr int BLOCK_ROWS = 8;
+
+template <typename T>
+__global__ void _Transpose3DKernel(const TArray<int64_t> input_shape,
+                                   const TArray<int64_t> input_strides,
+                                   const T* input_data, T* output_data) {
+  __shared__ T tile[TILE_DIM * (TILE_DIM+1)];
+
+  int x = blockIdx.x * TILE_DIM + threadIdx.x;
+  int y = blockIdx.y * TILE_DIM + threadIdx.y;
+
+  for (int j = 0; j < TILE_DIM; j += BLOCK_ROWS)
+    tile[(threadIdx.y+j) * TILE_DIM + threadIdx.x] = input_data[blockIdx.z * input_strides[0] + (y+j)*input_shape[2] + x];
+
+  __syncthreads();
+
+  x = blockIdx.y * TILE_DIM + threadIdx.x;
+  y = blockIdx.x * TILE_DIM + threadIdx.y;
+
+  for (int j = 0; j < TILE_DIM; j += BLOCK_ROWS)
+    output_data[blockIdx.z * input_strides[0] + (y+j)*input_shape[1] + x] = tile[threadIdx.x * TILE_DIM + threadIdx.y + j];
+}
+
+bool canDoTranspose3D(int32_t rank,
+                      const std::vector<int64_t>& input_dims,
+                      const std::vector<size_t>& permutations) {
+  if (rank == 3 &&
+      input_dims[1] % TILE_DIM == 0 && input_dims[2] % TILE_DIM == 0 &&
+      permutations[1] == 2 && permutations[2] == 1) {
+    return true;
+  }
+  return false;
+}
+
+Status Transpose3DImpl(size_t element_size,
+                       const TArray<int64_t>& input_shape, const TArray<int64_t>& input_strides,
+                       const void* input_data, void* output_data, int64_t N) {
+  dim3 block_size(TILE_DIM, BLOCK_ROWS);
+  dim3 grid_size(input_shape[2]/TILE_DIM, input_shape[1]/TILE_DIM, input_shape[0]);
+
+  switch (element_size) {
+    case sizeof(int8_t):
+      _Transpose3DKernel<int8_t><<<grid_size, block_size, 0>>>(
+          input_shape, input_strides,
+          reinterpret_cast<const ToCudaType<int8_t>::MappedType*>(input_data),
+          reinterpret_cast<ToCudaType<int8_t>::MappedType*>(output_data));
+      break;
+    case sizeof(int16_t):
+      _Transpose3DKernel<int16_t><<<grid_size, block_size, 0>>>(
+          input_shape, input_strides,
+          reinterpret_cast<const int16_t*>(input_data),
+          reinterpret_cast<int16_t*>(output_data));
+      break;
+    case sizeof(int32_t):
+      _Transpose3DKernel<int32_t><<<grid_size, block_size, 0>>>(
+          input_shape, input_strides,
+          reinterpret_cast<const ToCudaType<int32_t>::MappedType*>(input_data),
+          reinterpret_cast<ToCudaType<int32_t>::MappedType*>(output_data));
+      break;
+    case sizeof(int64_t):
+      _Transpose3DKernel<int64_t><<<grid_size, block_size, 0>>>(
+          input_shape, input_strides,
+          reinterpret_cast<const ToCudaType<int64_t>::MappedType*>(input_data),
+          reinterpret_cast<ToCudaType<int64_t>::MappedType*>(output_data));
+      break;
+    default:
+      return ORT_MAKE_STATUS(ONNXRUNTIME, FAIL, "Type not supported for transpose on CUDA. Element size was ",
+                              element_size);
+  }
+
+  return Status::OK();
+}
+
 template <typename T, int element_size>
 __global__ void _Transpose4DKernel(const TArray<int64_t> input_strides, const T* input_data,
                                    const TArray<int64_t> output_strides, T* output_data, CUDA_LONG N) {
@@ -27,6 +101,22 @@ __global__ void _Transpose4DKernel(const TArray<int64_t> input_strides, const T*
   if (input_index < N && output_index < N) {
     v_output[output_index] = v_input[input_index];
   }
+}
+
+bool canDoTranspose4D(const cudaDeviceProp& prop,
+                      size_t element_size,
+                      int32_t rank,
+                      const std::vector<int64_t>& input_dims,
+                      const std::vector<size_t>& permutations) {
+  // In Tranpose4D, vector of int4 will be used.
+  int64_t num_threads_per_block = input_dims[2] * input_dims[3] / (4 * sizeof(int) / element_size);
+  if (rank == 4 && permutations[3] == 3 &&
+      num_threads_per_block <= prop.maxThreadsPerBlock &&
+      // num_threads_per_block must be aligned with warp size: 32
+      ((num_threads_per_block & (prop.warpSize - 1)) == 0)) {
+    return true;
+  }
+  return false;
 }
 
 Status Transpose4DImpl(size_t element_size, const TArray<int64_t>& input_shape, const TArray<int64_t>& input_strides, const void* input_data,