In this PR, the following things are done to improve CUDA transpose kernel. With this PR, the throughput of BERT-L is improved by ~1.4%.

1.  Flatten the adjacent dimensions which are contiguous. For example: permutations[0, 2, 3, 1] -> [0, 2, 1], permutations[0, 3, 1, 2] -> [0, 2, 1]. This will make the loop iterations in transpose kernel less.
2.  For 3D tensor, if the permutations are done in last two dimensions, then we can easily use shared memory to improve GPU memory utilization.
3. For 4D tensor, if the permutations are not done in the last dimension, then we can use int4 to access memory. Besides, if the last two dimensions can fit in block size, then the kernel can be simplified by configuring block size with the shape of last two dimensions.

The table below is the improvement for BERT-L tensor with shape [64, 128, 16, 64]

  | Baseline (micro seconds) | Optimization (micro seconds) | Improvement
-- | -- | -- | --
Permutation: [0 2 3 1] | 110.531 | 95.808 | 13%
Permutation: [0 3 1 2] | 187.981 | 89.6352 | 52%
Permutation: [0 2 1 3] | 87.7408 | 31.0208 | 65%