@edgchen1 2020-06-29
yes, some UTs would be good. e.g., for the dimension flattening cases

---
In reply to: [644569847](https://github.com/microsoft/onnxruntime/pull/4233#issuecomment-644569847) [](ancestors = 644569847)